refactors and starts working on some sprite stuff
