+++ Assets/Materials/Physics/NoFriction.physicMaterial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fe7a3f39517fd3e4abcc78ba377445f8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 13400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
