+++ Assets/Materials/Render.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d73c3489f0897cd41acf6bbc9da7a6c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
