+++ Assets/Prefabs/Player.prefab
@@ -91,7 +91,7 @@ BoxCollider:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3715997352084929724}
-  m_Material: {fileID: 0}
+  m_Material: {fileID: 13400000, guid: fe7a3f39517fd3e4abcc78ba377445f8, type: 2}
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 2
@@ -129,3 +129,4 @@ MonoBehaviour:
   speed: 450
   rb: {fileID: 3715997352084929723}
   velocity: {x: 0, y: 0, z: 0}
+  health: 100
