+++ Assets/Scripts/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb6e4c0fe2aaa204aa2aac63def3de64
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
