+++ Assets/Scripts/Enemies/FollowPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e4bd6ce97683bf84bb5ed95403996054
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
