+++ Assets/Sprites/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60e109e25287a414792e100b177abee4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
